requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchvision==0.16.0
xformers
accelerate==0.24.0
# StreamDiffusion
fire==0.5.0
diffusers==0.24.0
omegaconf==2.3.0
cuda-python==12.3.0
onnx==1.15.0
onnxruntime==1.16.3
protobuf==3.20.2
colored==2.2.4
# streamdiffusion[tensorrt]
nvidia-cudnn-cu12==8.9.4.25
--extra-index-url https://pypi.nvidia.com
tensorrt==9.0.1.post11.dev4
--extra-index-url https://pypi.ngc.nvidia.com
polygraphy==0.47.1
onnx-graphsurgeon==0.3.26
# PIA
decord==0.6.0
einops==0.7.0
# PTI
lpips==0.1.4
wandb==0.16.2
matplotlib==3.8.2
dlib==19.24.2
# DragGAN
scipy==1.11.4
Ninja==1.10.2
imageio-ffmpeg==0.4.3
huggingface-hub==0.20.1
hf-transfer==0.1.4
pillow==9.5.0
imageio==2.33.1
# GUI requirements
pyqt5==5.15.10  # Local Demo
gradio==3.44.0  # Web Demo
# others
opencv-python-headless==4.9.0.80
transformers==4.35.2
moviepy==1.0.3
scikit-image==0.22.0
ipython==8.20.0
gdown==5.0
termcolor==2.4.0
importlib-metadata==7.0.1